user pins, problem reporting via flags
